Set Older Project Data under Version Control
In This Commit, I Set up the version control for the DotAutomatedClassDiagramm Project.
At this state of the project, the program is able to read out C# class-files from a Project
and generate a simple DotClassDiagramm with all found class names. Furthermore, it also
generates a PNG File as result.