Re structuring of dotClassContainer parsing
Features
- creating list of dotClassContainers
- build up diagram with classContainer entrys
- add git ignore file for vscode

Details
In order to speed up the process of parsing and have a clear separation of parsing and diagram creation,
the program fills now a list of dotClassContainer with parsed class- and attribute names. With the information
stored in the list, the diagram is created in the next step.

Refferenz

append c# lists
https://www.tutorialspoint.com/How-to-append-a-second-list-to-an-existing-list-in-Chash