Definition of Input Arguments for the Conversion
Features:
- Input Argument description
- Bugfix destination- and sourcepath
- relocation of  DotClassDiagrammHead.txt to resources Folder
- Creation of a ContextClass that handles the input Arguments with the help of NDesk.Options

Details:
In order to give the user an interface to change the path in- or output, the program has now argument options.
In Futre Changes it shell also be possible to alter the programming language and an option to
generate a classdiagramm from Code or create code from dotClassdiagramm.